Refactor service imports in routers and services
